Jetpack Licensing: Streamlined logic for license issuing and payment method requirements (#65901)
* Move license issuing mutation definition to a hook

* Add callback to the hook to require payment method

* Add redirect for payment method requiriment

* Remove creating licenses from GET requests

* Fix typo

* Refactor to issue license when adding payment method

* Refactor to extract product state from hook to component

* Add conditional href to payment method addition back button

* Refactor to enable license assingment payment method requirement

* Move license issuing auxiliary funtions to `partner-portal/utils`;

* Update payment-method-add tests

* Refactor license issuing, cleanup of old logic

* Refactor to improve component logic and tracking

* Fix payment method check middleware wrong redirect

Valid payment method check middleware was redirecting not to the page
that the user was navigating to, but to the one they were at.

* Refactor license issue hook for simpler logic

* Refactor license issuing form to work w/ new hook

* Rename utilitty function for clearer understanding

* Refactor payment method add form removing unecessary logic

* Remove overkill memoization

* Update license issuing tracked event name

* Update moving tracking from component to hook level

* Fix conditionals on successful payment method addition

* Add notice informing which license was issued

* Refactor success of payment method addition to resunme the flow

* Revert changes due no longer reproducible bug

There was a bug that was redirecting the user to payment methods
but that seemed to be fixed with other changes
